+++ RELEASES.rst
@@ -1,4 +1,6 @@
 Releases
 ========
 
+* TBD:
+    * Allow specifying docker-compose build service
 * 0.0.1, 2018-11-07: initial release